grass.gunittest: Place TestResult in result.py like Python 3.13
