+++ python/grass/gunittest/Makefile
@@ -8,7 +8,7 @@ GDIR = $(PYDIR)/grass
 DSTDIR = $(GDIR)/gunittest
 
 # TODO: add multireport multirunner
-MODULES = case gmodules loader runner checkers gutils invoker main reporters utils
+MODULES = case gmodules loader runner checkers gutils invoker main reporters result utils
 
 PYFILES := $(patsubst %,$(DSTDIR)/%.py,$(MODULES) __init__)
 PYCFILES := $(patsubst %,$(DSTDIR)/%.pyc,$(MODULES) __init__)
