+++ python/grass/gunittest/runner.py
@@ -16,6 +16,7 @@ import sys
 import time
 
 import unittest
+from grass.gunittest.result import TestResult
 
 __unittest = True
 
@@ -35,28 +36,6 @@ class _WritelnDecorator:
         if arg:
             self.write(arg)
         self.write("\n")  # text-mode streams translate to \r\n if needed
-
-
-class TestResult(unittest.TestResult):
-    # descriptions and verbosity unused
-    # included for compatibility with unittest's TestResult
-    # where are also unused, so perhaps we can remove them
-    # stream set to None and not included in interface, it would not make sense
-    def __init__(self, stream=None, descriptions=None, verbosity=None):
-        super().__init__(stream=stream, descriptions=descriptions, verbosity=verbosity)
-        self.successes = []
-
-    def addSuccess(self, test):
-        super().addSuccess(test)
-        self.successes.append(test)
-
-    # TODO: better would be to pass start at the beginning
-    # alternative is to leave counting time on class
-    # TODO: document: we expect all grass classes to have setTimes
-    # TODO: alternatively, be more forgiving for non-unittest methods
-    def setTimes(self, start_time, end_time, time_taken):
-        pass
-        # TODO: implement this
 
 
 class TextTestResult(TestResult):
